finish

+++ src/big5Hardware/OutputDevice.java
@@ -1,7 +1,7 @@
 package src.big5Hardware;
 
 public class OutputDevice {
-    public void displayResult(int result) {
+    public void displayResult(double  result) {
         System.out.println("演算結果は" + result + "です");
     }
